+++ src/config.ts
@@ -108,7 +108,7 @@ export interface ExtensionConfigurationSettings {
     buildToolArgs: string[];
     parallelJobs: number | undefined;
     ctestPath: string;
-    ctest: { parallelJobs: number };
+    ctest: { parallelJobs: number; allowParallelJobs: boolean };
     parseBuildDiagnostics: boolean;
     enabledOutputParsers: string[];
     debugConfig: CppDebugConfiguration;
@@ -301,6 +301,9 @@ export class ConfigurationReader implements vscode.Disposable {
     get ctestParallelJobs(): number | null {
         return this.configData.ctest.parallelJobs;
     }
+    get ctestAllowParallelJobs(): boolean {
+        return this.configData.ctest.allowParallelJobs;
+    }
     get parseBuildDiagnostics(): boolean {
         return !!this.configData.parseBuildDiagnostics;
     }
@@ -484,7 +487,7 @@ export class ConfigurationReader implements vscode.Disposable {
         buildToolArgs: new vscode.EventEmitter<string[]>(),
         parallelJobs: new vscode.EventEmitter<number>(),
         ctestPath: new vscode.EventEmitter<string>(),
-        ctest: new vscode.EventEmitter<{ parallelJobs: number }>(),
+        ctest: new vscode.EventEmitter<{ parallelJobs: number; allowParallelJobs: boolean }>(),
         parseBuildDiagnostics: new vscode.EventEmitter<boolean>(),
         enabledOutputParsers: new vscode.EventEmitter<string[]>(),
         debugConfig: new vscode.EventEmitter<CppDebugConfiguration>(),