+++ src/extension.ts
@@ -30,7 +30,7 @@ import rollbar from '@cmt/rollbar';
 import { StateManager } from './state';
 import { cmakeTaskProvider, CMakeTaskProvider } from '@cmt/cmakeTaskProvider';
 import * as telemetry from '@cmt/telemetry';
-import { ProjectOutline, TargetNode, SourceFileNode, WorkspaceFolderNode } from '@cmt/projectOutline';
+import { ProjectOutline, ProjectNode, TargetNode, SourceFileNode, WorkspaceFolderNode } from '@cmt/projectOutline';
 import * as util from '@cmt/util';
 import { ProgressHandle, DummyDisposable, reportProgress, runCommand } from '@cmt/util';
 import { DEFAULT_VARIANTS } from '@cmt/variant';
@@ -45,6 +45,7 @@ import { StatusBar } from '@cmt/status';
 import { DebugAdapterNamedPipeServerDescriptorFactory } from './debug/debugAdapterNamedPipeServerDescriptorFactory';
 import { getCMakeExecutableInformation } from './cmake/cmakeExecutable';
 import { DebuggerInformation, getDebuggerPipeName } from './debug/debuggerConfigureDriver';
+import { DebugConfigurationProvider, DynamicDebugConfigurationProvider } from './debug/debugConfigurationProvider';
 
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
 const localize: nls.LocalizeFunc = nls.loadMessageBundle();
@@ -487,31 +488,31 @@ export class ExtensionManager implements vscode.Disposable {
                 shouldConfigure = chosen.doConfigure;
             }
         }
-        if (project) {
-            if (!project.hasCMakeLists()) {
+        if (!project.hasCMakeLists()) {
+            if (shouldConfigure === true) {
                 await project.cmakePreConditionProblemHandler(CMakePreconditionProblems.MissingCMakeListsFile, false, this.workspaceConfig);
+            }
+        } else {
+            if (shouldConfigure === true) {
+                // We've opened a new workspace folder, and the user wants us to
+                // configure it now.
+                log.debug(localize('configuring.workspace.on.open', 'Configuring workspace on open {0}', project.folderPath));
+                await this.configureExtensionInternal(ConfigureTrigger.configureOnOpen, project);
             } else {
-                if (shouldConfigure === true) {
-                    // We've opened a new workspace folder, and the user wants us to
-                    // configure it now.
-                    log.debug(localize('configuring.workspace.on.open', 'Configuring workspace on open {0}', project.folderPath));
-                    await this.configureExtensionInternal(ConfigureTrigger.configureOnOpen, project);
+                const configureButtonMessage = localize('configure.now.button', 'Configure Now');
+                let result: string | undefined;
+                if (silentScanForKitsNeeded) {
+                    // This popup will show up the first time after deciding not to configure, if a version change has been detected
+                    // in the kits definition. This may happen during a CMake Tools extension upgrade.
+                    // The warning is emitted only once because scanForKitsIfNeeded returns true only once after such change,
+                    // being tied to a global state variable.
+                    result = await vscode.window.showWarningMessage(localize('configure.recommended', 'It is recommended to reconfigure after upgrading to a new kits definition.'), configureButtonMessage);
+                }
+                if (result === configureButtonMessage) {
+                    await this.configureExtensionInternal(ConfigureTrigger.buttonNewKitsDefinition, project);
                 } else {
-                    const configureButtonMessage = localize('configure.now.button', 'Configure Now');
-                    let result: string | undefined;
-                    if (silentScanForKitsNeeded) {
-                        // This popup will show up the first time after deciding not to configure, if a version change has been detected
-                        // in the kits definition. This may happen during a CMake Tools extension upgrade.
-                        // The warning is emitted only once because scanForKitsIfNeeded returns true only once after such change,
-                        // being tied to a global state variable.
-                        result = await vscode.window.showWarningMessage(localize('configure.recommended', 'It is recommended to reconfigure after upgrading to a new kits definition.'), configureButtonMessage);
-                    }
-                    if (result === configureButtonMessage) {
-                        await this.configureExtensionInternal(ConfigureTrigger.buttonNewKitsDefinition, project);
-                    } else {
-                        log.debug(localize('using.cache.to.configure.workspace.on.open', 'Attempting to use cache to configure workspace {0}', rootFolder.uri.toString()));
-                        await this.configureExtensionInternal(ConfigureTrigger.configureWithCache, project);
-                    }
+                    log.debug(localize('using.cache.to.configure.workspace.on.open', 'Attempting to use cache to configure workspace {0}', rootFolder.uri.toString()));
+                    await this.configureExtensionInternal(ConfigureTrigger.configureWithCache, project);
                 }
             }
         }
@@ -628,14 +629,7 @@ export class ExtensionManager implements vscode.Disposable {
             return;
         }
         const folder: vscode.WorkspaceFolder = cmakeProject.workspaceFolder;
-        this.projectOutline.updateCodeModel(
-            cmakeProject.workspaceContext.folder,
-            cmakeProject.codeModelContent,
-            {
-                defaultTarget: cmakeProject.defaultBuildTarget || undefined,
-                launchTargetName: cmakeProject.launchTargetName
-            }
-        );
+        this.projectOutline.updateCodeModel(cmakeProject, cmakeProject.codeModelContent);
         rollbar.invokeAsync(localize('update.code.model.for.cpptools', 'Update code model for cpptools'), {}, async () => {
             if (vscode.workspace.getConfiguration('C_Cpp', folder.uri).get<string>('intelliSenseEngine')?.toLocaleLowerCase() === 'disabled') {
                 log.debug(localize('update.intellisense.disabled', 'Not updating the configuration provider because {0} is set to {1}', '"C_Cpp.intelliSenseEngine"', '"Disabled"'));
@@ -677,14 +671,7 @@ export class ExtensionManager implements vscode.Disposable {
                 } else if (drv && drv.codeModelContent) {
                     codeModelContent = drv.codeModelContent;
                     this.configProvider.updateConfigurationData({ cache, codeModelContent, clCompilerPath, activeTarget: cmakeProject.defaultBuildTarget, activeBuildTypeVariant: actualBuildType, folder: cmakeProject.folderPath });
-                    this.projectOutline.updateCodeModel(
-                        cmakeProject.workspaceContext.folder,
-                        codeModelContent,
-                        {
-                            defaultTarget: cmakeProject.defaultBuildTarget || undefined,
-                            launchTargetName: cmakeProject.launchTargetName
-                        }
-                    );
+                    this.projectOutline.updateCodeModel(cmakeProject, codeModelContent);
                 }
                 // Inform cpptools that custom CppConfigurationProvider will be able to service the current workspace.
                 this.ensureCppToolsProviderRegistered();
@@ -1044,7 +1031,7 @@ export class ExtensionManager implements vscode.Disposable {
         return 0;
     }
 
-    private getProjectFromFolder(folder?: vscode.WorkspaceFolder | string) {
+    private getProjectFromFolder(folder?: vscode.WorkspaceFolder | string, sourceDir?: string) {
         const workspaceFolder: vscode.WorkspaceFolder | undefined = this.getWorkspaceFolder(folder);
         if (workspaceFolder) {
             const activeProject: CMakeProject | undefined = this.getActiveProject();
@@ -1052,24 +1039,22 @@ export class ExtensionManager implements vscode.Disposable {
             if (!projects || projects.length === 0) {
                 return activeProject;
             } else {
-                for (const project of projects) {
-                    // Choose the active project.
-                    if (activeProject?.folderPath === project.folderPath) {
-                        return project;
-                    }
-                }
-                // Choose the first project folder.
-                return projects[0];
+                // Choose project by corresponding source directory
+                return projects.find(project => sourceDir && (path.normalize(sourceDir) === path.normalize(project.folderPath)))
+                    // Choose project by folder of active project
+                    ?? projects.find(project => activeProject?.folderPath === project.folderPath)
+                    // Fallback to first project
+                    ?? projects[0];
             }
         }
         return undefined;
     }
 
-    runCMakeCommand(command: RunCMakeCommand, folder?: vscode.WorkspaceFolder, precheck?: (cmakeProject: CMakeProject) => Promise<boolean>, cleanOutputChannel?: boolean): Promise<any> {
+    runCMakeCommand(command: RunCMakeCommand, folder?: vscode.WorkspaceFolder, precheck?: (cmakeProject: CMakeProject) => Promise<boolean>, cleanOutputChannel?: boolean, sourceDir?: string): Promise<any> {
         if (cleanOutputChannel) {
             this.cleanOutputChannel();
         }
-        const project = this.getProjectFromFolder(folder);
+        const project = this.getProjectFromFolder(folder, sourceDir);
         if (project) {
             return this.runCMakeCommandForProject(command, project, precheck);
         }
@@ -1116,11 +1101,11 @@ export class ExtensionManager implements vscode.Disposable {
         return this.runCMakeCommandForAll(cmakeProject => cmakeProject.cleanConfigureWithDebugger(ConfigureTrigger.commandCleanConfigureAllWithDebugger, debuggerInformation), undefined, true);
     }
 
-    configure(folder?: vscode.WorkspaceFolder, showCommandOnly?: boolean) {
+    configure(folder?: vscode.WorkspaceFolder, showCommandOnly?: boolean, sourceDir?: string) {
         telemetry.logEvent("configure", { all: "false", debug: "false"});
         return this.runCMakeCommand(
             cmakeProject => cmakeProject.configureInternal(ConfigureTrigger.commandConfigure, [], showCommandOnly ? ConfigureType.ShowCommandOnly : ConfigureType.Normal),
-            folder, undefined, true);
+            folder, undefined, true, sourceDir);
     }
 
     configureWithDebugger(folder?: vscode.WorkspaceFolder) {
@@ -1135,7 +1120,7 @@ export class ExtensionManager implements vscode.Disposable {
     }
 
     showConfigureCommand(folder?: vscode.WorkspaceFolder) {
-        return this.configure(folder, true);
+        return this.configure(folder, true, undefined);
     }
 
     configureAll() {
@@ -1157,13 +1142,20 @@ export class ExtensionManager implements vscode.Disposable {
         return this.runCMakeCommand(cmakeProject => cmakeProject.editCacheUI());
     }
 
-    build(folder?: vscode.WorkspaceFolder, name?: string, showCommandOnly?: boolean, isBuildCommand?: boolean) {
+    build(folder?: vscode.WorkspaceFolder, name?: string, sourceDir?: string, showCommandOnly?: boolean, isBuildCommand?: boolean) {
         telemetry.logEvent("build", { all: "false"});
-        return this.runCMakeCommand(cmakeProject => cmakeProject.build(name ? [name] : undefined, showCommandOnly, (isBuildCommand === undefined) ? true : isBuildCommand), folder, this.ensureActiveBuildPreset, true);
+        return this.runCMakeCommand(cmakeProject => {
+            const targets = name ? [name] : undefined;
+            return cmakeProject.build(targets, showCommandOnly, (isBuildCommand === undefined) ? true : isBuildCommand);
+        },
+        folder,
+        this.ensureActiveBuildPreset,
+        true,
+        sourceDir);
     }
 
     showBuildCommand(folder?: vscode.WorkspaceFolder, name?: string) {
-        return this.build(folder, name, true, false);
+        return this.build(folder, name, undefined, true, false);
     }
 
     buildAll(name?: string | string[]) {
@@ -1176,8 +1168,8 @@ export class ExtensionManager implements vscode.Disposable {
         true);
     }
 
-    setDefaultTarget(folder?: vscode.WorkspaceFolder, name?: string) {
-        return this.runCMakeCommand(cmakeProject => cmakeProject.setDefaultTarget(name), folder);
+    setDefaultTarget(folder?: vscode.WorkspaceFolder, name?: string, sourceDir?: string) {
+        return this.runCMakeCommand(cmakeProject => cmakeProject.setDefaultTarget(name), folder, undefined, undefined, sourceDir);
     }
 
     setVariant(folder?: vscode.WorkspaceFolder, name?: string) {
@@ -1218,7 +1210,7 @@ export class ExtensionManager implements vscode.Disposable {
 
     clean(folder?: vscode.WorkspaceFolder) {
         telemetry.logEvent("clean", { all: "false"});
-        return this.build(folder, 'clean', undefined, false);
+        return this.build(folder, 'clean', undefined, undefined, false);
     }
 
     cleanAll() {
@@ -1341,34 +1333,70 @@ export class ExtensionManager implements vscode.Disposable {
         return this.runCMakeCommandForProject(cmakeProject => cmakeProject.quickStart(folder));
     }
 
-    launchTargetPath(folder?: vscode.WorkspaceFolder | string) {
+    launchTargetPath({ folder, targetName }: { folder?: vscode.WorkspaceFolder | string; targetName?: string } = {}) {
         telemetry.logEvent("substitution", { command: "launchTargetPath" });
-        return this.queryCMakeProject(cmakeProject => cmakeProject.launchTargetPath(), folder);
+        return this.queryCMakeProject(async cmakeProject => {
+            if (targetName !== undefined && targetName !== null) {
+                await cmakeProject.setLaunchTargetByName(targetName);
+            }
+            const targetPath = await cmakeProject.launchTargetPath();
+            return targetPath;
+        }, folder);
     }
 
-    launchTargetDirectory(folder?: vscode.WorkspaceFolder | string) {
+    launchTargetDirectory({ folder, targetName }: { folder?: vscode.WorkspaceFolder | string; targetName?: string } = {}) {
         telemetry.logEvent("substitution", { command: "launchTargetDirectory" });
-        return this.queryCMakeProject(cmakeProject => cmakeProject.launchTargetDirectory(), folder);
+        return this.queryCMakeProject(async cmakeProject => {
+            if (targetName !== undefined && targetName !== null) {
+                await cmakeProject.setLaunchTargetByName(targetName);
+            }
+            const targetDirectory = await cmakeProject.launchTargetDirectory();
+            return targetDirectory;
+        }, folder);
     }
 
-    launchTargetFilename(folder?: vscode.WorkspaceFolder | string) {
+    launchTargetFilename({ folder, targetName }: { folder?: vscode.WorkspaceFolder | string; targetName?: string } = {}) {
         telemetry.logEvent("substitution", { command: "launchTargetFilename" });
-        return this.queryCMakeProject(cmakeProject => cmakeProject.launchTargetFilename(), folder);
+        return this.queryCMakeProject(async cmakeProject => {
+            if (targetName !== undefined && targetName !== null) {
+                await cmakeProject.setLaunchTargetByName(targetName);
+            }
+            const targetFilename = await cmakeProject.launchTargetFilename();
+            return targetFilename;
+        }, folder);
     }
 
-    getLaunchTargetPath(folder?: vscode.WorkspaceFolder | string) {
+    getLaunchTargetPath({ folder, targetName }: { folder?: vscode.WorkspaceFolder | string; targetName?: string } = {}) {
         telemetry.logEvent("substitution", { command: "getLaunchTargetPath" });
-        return this.queryCMakeProject(cmakeProject => cmakeProject.getLaunchTargetPath(), folder);
+        return this.queryCMakeProject(async cmakeProject => {
+            if (targetName !== undefined && targetName !== null) {
+                await cmakeProject.setLaunchTargetByName(targetName);
+            }
+            const targetPath = await cmakeProject.getLaunchTargetPath();
+            return targetPath;
+        }, folder);
     }
 
-    getLaunchTargetDirectory(folder?: vscode.WorkspaceFolder | string) {
+    getLaunchTargetDirectory({ folder, targetName }: { folder?: vscode.WorkspaceFolder | string; targetName?: string } = {}) {
         telemetry.logEvent("substitution", { command: "getLaunchTargetDirectory" });
-        return this.queryCMakeProject(cmakeProject => cmakeProject.getLaunchTargetDirectory(), folder);
+        return this.queryCMakeProject(async cmakeProject => {
+            if (targetName !== undefined && targetName !== null) {
+                await cmakeProject.setLaunchTargetByName(targetName);
+            }
+            const targetDirectory = await cmakeProject.getLaunchTargetDirectory();
+            return targetDirectory;
+        }, folder);
     }
 
-    getLaunchTargetFilename(folder?: vscode.WorkspaceFolder | string) {
+    getLaunchTargetFilename({ folder, targetName }: { folder?: vscode.WorkspaceFolder | string; targetName?: string } = {}) {
         telemetry.logEvent("substitution", { command: "getLaunchTargetFilename" });
-        return this.queryCMakeProject(cmakeProject => cmakeProject.getLaunchTargetFilename(), folder);
+        return this.queryCMakeProject(async cmakeProject => {
+            if (targetName !== undefined && targetName !== null) {
+                await cmakeProject.setLaunchTargetByName(targetName);
+            }
+            const targetFilename = await cmakeProject.getLaunchTargetFilename();
+            return targetFilename;
+        }, folder);
     }
 
     buildTargetName(folder?: vscode.WorkspaceFolder | string) {
@@ -1401,9 +1429,9 @@ export class ExtensionManager implements vscode.Disposable {
         return this.queryCMakeProject(cmakeProject => cmakeProject.tasksBuildCommand(), folder);
     }
 
-    debugTarget(folder?: vscode.WorkspaceFolder, name?: string): Promise<vscode.DebugSession | null> {
+    debugTarget(folder?: vscode.WorkspaceFolder, name?: string, sourceDir?: string): Promise<vscode.DebugSession | null> {
         telemetry.logEvent("debug", { all: "false" });
-        return this.runCMakeCommand(cmakeProject => cmakeProject.debugTarget(name), folder);
+        return this.runCMakeCommand(cmakeProject => cmakeProject.debugTarget(name), folder, undefined, undefined, sourceDir);
     }
 
     async debugTargetAll(): Promise<(vscode.DebugSession | null)[]> {
@@ -1415,9 +1443,9 @@ export class ExtensionManager implements vscode.Disposable {
         return debugSessions;
     }
 
-    launchTarget(folder?: vscode.WorkspaceFolder, name?: string): Promise<vscode.Terminal | null> {
+    launchTarget(folder?: vscode.WorkspaceFolder, name?: string, sourceDir?: string): Promise<vscode.Terminal | null> {
         telemetry.logEvent("launch", { all: "false" });
-        return this.runCMakeCommand(cmakeProject => cmakeProject.launchTarget(name), folder);
+        return this.runCMakeCommand(cmakeProject => cmakeProject.launchTarget(name), folder, undefined, undefined, sourceDir);
     }
 
     async launchTargetAll(): Promise<(vscode.Terminal | null)[]> {
@@ -1429,8 +1457,8 @@ export class ExtensionManager implements vscode.Disposable {
         return terminals;
     }
 
-    selectLaunchTarget(folder?: vscode.WorkspaceFolder, name?: string) {
-        return this.runCMakeCommand(cmakeProject => cmakeProject.selectLaunchTarget(name), folder);
+    selectLaunchTarget(folder?: vscode.WorkspaceFolder, name?: string, sourceDir?: string) {
+        return this.runCMakeCommand(cmakeProject => cmakeProject.selectLaunchTarget(name), folder, undefined, undefined, sourceDir);
     }
 
     async resetState(folder?: vscode.WorkspaceFolder) {
@@ -1442,6 +1470,10 @@ export class ExtensionManager implements vscode.Disposable {
         }
 
         void vscode.commands.executeCommand('workbench.action.reloadWindow');
+    }
+
+    openSettings() {
+        void vscode.commands.executeCommand('workbench.action.openSettings', '@ext:ms-vscode.cmake-tools');
     }
 
     async viewLog() {
@@ -1731,6 +1763,14 @@ async function setup(context: vscode.ExtensionContext, progress?: ProgressHandle
         )
     );
 
+    context.subscriptions.push(vscode.debug.registerDebugConfigurationProvider("cmake", new DebugConfigurationProvider()));
+    context.subscriptions.push(
+        vscode.debug.registerDebugConfigurationProvider(
+            "cmake",
+            new DynamicDebugConfigurationProvider(),
+            vscode.DebugConfigurationProviderTriggerKind.Dynamic)
+    );
+
     // List of functions that will be bound commands
     const funs: (keyof ExtensionManager)[] = [
         'activeFolderName',
@@ -1804,6 +1844,7 @@ async function setup(context: vscode.ExtensionContext, progress?: ProgressHandle
         'selectLaunchTarget',
         'setDefaultTarget',
         'resetState',
+        'openSettings',
         'viewLog',
         'logDiagnostics',
         'compileFile',
@@ -1840,12 +1881,21 @@ async function setup(context: vscode.ExtensionContext, progress?: ProgressHandle
         vscode.commands.registerCommand('cmake.outline.editCacheUI', () => runCommand('editCacheUI')),
         vscode.commands.registerCommand('cmake.outline.cleanRebuildAll', () => runCommand('cleanRebuildAll')),
         // Commands for outline items
-        vscode.commands.registerCommand('cmake.outline.buildTarget', (what: TargetNode) => runCommand('build', what.folder, what.name)),
-        vscode.commands.registerCommand('cmake.outline.runUtilityTarget', (what: TargetNode) => runCommand('build', what.folder, what.name)),
-        vscode.commands.registerCommand('cmake.outline.debugTarget', (what: TargetNode) => runCommand('debugTarget', what.folder, what.name)),
-        vscode.commands.registerCommand('cmake.outline.launchTarget', (what: TargetNode) => runCommand('launchTarget', what.folder, what.name)),
-        vscode.commands.registerCommand('cmake.outline.setDefaultTarget', (what: TargetNode) => runCommand('setDefaultTarget', what.folder, what.name)),
-        vscode.commands.registerCommand('cmake.outline.setLaunchTarget', (what: TargetNode) => runCommand('selectLaunchTarget', what.folder, what.name)),
+        vscode.commands.registerCommand('cmake.outline.configure', async (what: ProjectNode|SourceFileNode) => {
+            if (what instanceof ProjectNode) {
+                await runCommand('configure', what.folder, false, what.sourceDirectory);
+            } else if (what instanceof SourceFileNode) {
+                await runCommand('configure', what.folder, false, what.sourcePath);
+            }
+        }),
+        vscode.commands.registerCommand('cmake.outline.build', (what: ProjectNode) => runCommand('build', what.folder, "all", what.sourceDirectory)),
+        vscode.commands.registerCommand('cmake.outline.clean', (what: ProjectNode) => runCommand('build', what.folder, "clean", what.sourceDirectory)),
+        vscode.commands.registerCommand('cmake.outline.buildTarget', (what: TargetNode) => runCommand('build', what.folder, what.name, what.sourceDir)),
+        vscode.commands.registerCommand('cmake.outline.runUtilityTarget', (what: TargetNode) => runCommand('build', what.folder, what.name, what.sourceDir)),
+        vscode.commands.registerCommand('cmake.outline.debugTarget', (what: TargetNode) => runCommand('debugTarget', what.folder, what.name, what.sourceDir)),
+        vscode.commands.registerCommand('cmake.outline.launchTarget', (what: TargetNode) => runCommand('launchTarget', what.folder, what.name, what.sourceDir)),
+        vscode.commands.registerCommand('cmake.outline.setDefaultTarget', (what: TargetNode) => runCommand('setDefaultTarget', what.folder, what.name, what.sourceDir)),
+        vscode.commands.registerCommand('cmake.outline.setLaunchTarget', (what: TargetNode) => runCommand('selectLaunchTarget', what.folder, what.name, what.sourceDir)),
         vscode.commands.registerCommand('cmake.outline.revealInCMakeLists', (what: TargetNode) => what.openInCMakeLists()),
         vscode.commands.registerCommand('cmake.outline.compileFile', (what: SourceFileNode) => runCommand('compileFile', what.filePath)),
         // vscode.commands.registerCommand('cmake.outline.selectWorkspace', (what: WorkspaceFolderNode) => runCommand('selectWorkspace', what.wsFolder))